# Some arithmetic improvements