```
> go test -run Perf

Before:
  mul 597.14µs
  sqr 331.221µs
  add 90.095µs
  inv 14.161627ms

After:
  mul 404.358µs
  sqr 230.457µs
  add 75.695µs
  inv 11.213777ms
```